requirements.txt
streamlit
numpy
pandas
pillow
streamlit-image-coordinates
opencv-python